phase-2-production/requirements.txt
# Phase 2 Production: Chunking, Embedding, Indexing, Query API
# Python 3.10+

# Core dependencies
python-dotenv>=1.0.0

# NLP & Chunking
nltk>=3.8.1
beautifulsoup4>=4.12.0
transformers>=4.30.0
tokenizers>=0.13.3

# Embeddings
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Vector Databases
pinecone-client>=2.2.0
# Alternative: weaviate-client>=3.15.0

# Keyword Indexing
whoosh>=2.7.4

# FastAPI & Web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Caching
redis>=5.0.0

# Async & Retry Logic
tenacity>=8.2.0
aiohttp>=3.8.0

# Database (for metadata/logging)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Monitoring & Logging
python-json-logger>=2.0.7

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# LLM Clients (conditional)
openai>=1.0.0  # For GPT-4
anthropic>=0.7.0  # For Claude 3.5
# llama-cpp-python>=0.2.0  # For Llama 2 local
